opt code organization
